Rewrite to use single structure to hold session
The `ob-tmux-' object holds the information for the current
session. This makes it easier to shuttle more information between
functions. For instance, implementing socket support is going to be a
lot easier now. Validating the session name is going to be easier too,
because there is a single point where it can fail.

All internal functions have been renamed to use the `ob-tmux--'
prefix. This makes the code easier to write and shorter.

The idea for the cl-defstruct came from:
    https://nullprogram.com/blog/2018/02/14/

In Emacs 26, there should be even more support for this kind of
structure definition.