Fix issue when creating session with non-default window name
Before this change if I executed a code block like the following

  #+begin_src tmux :session default:foo
    echo hello world
  #+end_src

and no session named "default" exists, then first a session with a
window named "foo" would be created by (ob-tmux--create-session) and
then (sometimes) an additional window with the name "foo" would be
created by (ob-tmux--create-window).  This has lead in my case to some
follow up failures.  Therefore this change where always a window with
the default name is created for new sessions.