Kill default session before creating new one
In the `org-babel-execute:tmux' function, if the current session is a default
session, kill it before creating a new one. This ensures that a new session is
always created, even if a default session already exists.

This is necessary for safety concerns, as reusing the default session may result
in unexpected behavior due to leftover variables from previous sessions.

- Check if current session is default session
- Kill default session before creating new one